Batch submodule file digests in changeset path via shared gitbatch package
The changeset digest path was shelling out one git-show per submodule
file, reintroducing the 1-fork-per-file fan-out that the batched
cat-file helpers were created to avoid.

Move the batch implementation into a new internal/gitbatch package so
both controlservice and repositorychangeset can use it. The worktree
variants (TreeEntriesForFilesInWorktree, FileDigestsInWorktree) are now
used in digestPathsFromBase: submodule paths are grouped by worktree,
enumerated with a single ls-files --stage call and digested with a
single cat-file --batch session per submodule rather than per file.

controlservice/git_batch.go becomes thin wrappers that delegate to
gitbatch, keeping the existing controlservice tests unchanged.
